Issue #23: added the HTML_MINIFY flag setting
django-htmlmin middleware now supports the HTML_MINIFY setting. The
default value for the setting is not DEBUG. It means that if you don't
set HTML_MINIFY value, django-htmlmin will minify the response only
when DEBUG is off.

Examples of settings:

    DEBUG = True # will not minify

    ----

    DEBUG = True
    HTML_MINIFY = True # will minify

    ----

    DEBUG = False # will minify

    ----

    DEBUG = False
    HTML_MINIFY = False # will not minify

The issue will be closed when this change is available for the final
user through PyPI.