Issue #24: stopped always adding DOCTYPE
the html_minify function always added the doctype to the minified HTML,
even if it was not present in the original HTML. Now the DOCTYPE is
present in the minified HTML when it is present in the original HTML.

Also added a test for non-HTML5-two-lines doctype (we didn't have one).

This change will be included in the next django-htmlmin release, and the
we will be able to close the issue.