Fix test name (since the underlying behavior changed)
The new behavior is correct according to HTML5 parsing and serialization
rules. A <textarea> can not contain other elements, if it does, that is
an error and the result should try to match the user’s intent. So, like
most browsers, HTML inside a <textarea> is properly escaped upon
parsing.